Move new action picker to oss folder
