--- enterprise/frontend/src/metabase-enterprise/table-actions/settings/ActionPicker/index.ts
@@ -1,2 +1,0 @@
-export { TableActionPicker } from "./TableActionPicker";
-export { ModelActionPicker } from "./ModelActionPicker";
