--- enterprise/frontend/src/metabase-enterprise/table-actions/settings/ActionPicker/types.ts
@@ -1,5 +1,0 @@
-import type { CollectionItem } from "metabase/common/components/DataPicker";
-
-export type CollectionListItem = CollectionItem & {
-  position: number | null;
-};
