+++ frontend/src/metabase/common/components/DataPicker/types.ts
@@ -1,14 +1,11 @@
 import type {
   CardId,
   Collection,
-  CollectionId,
   DashboardId,
   DataGridWritebackActionId,
   DatabaseId,
   SchemaName,
-  TableActionId,
   TableId,
-  WritebackActionId,
 } from "metabase-types/api";
 
 import type { EntityPickerModalOptions } from "../EntityPicker";
@@ -71,23 +68,6 @@ export type ActionItem = {
   model: "action";
 };
 
-export type TableActionPickerValue = {
-  id: TableActionId;
-  name: string;
-  model: "action";
-  db_id: DatabaseId;
-  schema: SchemaName;
-  table_id: TableId;
-};
-
-export type ModelActionPickerValue = {
-  id: WritebackActionId;
-  name: string;
-  model: "action";
-  model_id: CardId;
-  collection_id?: CollectionId;
-};
-
 export type TablePickerValue = {
   id: TableId;
   name: string;
@@ -124,20 +104,3 @@ export type TablePickerStatePath = [
   SchemaName | undefined,
   TableId | undefined,
 ];
-
-export type TableActionPickerFolderItem = DatabaseItem | SchemaItem | TableItem;
-export type TableActionPickerStatePath = [
-  DatabaseId | undefined,
-  SchemaName | undefined,
-  TableId | undefined,
-  TableActionId | undefined,
-];
-export type TableActionPickerItem = TableActionPickerFolderItem | ActionItem;
-
-export type ModelActionPickerFolderItem = CollectionItem | ModelItem;
-export type ModelActionPickerStatePath = [
-  CollectionId | undefined,
-  CardId | undefined,
-  TableActionId | undefined,
-];
-export type ModelActionPickerItem = ModelActionPickerFolderItem | ActionItem;
