+++ frontend/src/metabase/common/components/TableOrModelActionPicker/index.ts
@@ -1,0 +1,1 @@
+export { TableOrModelActionPicker } from "./TableOrModelActionPicker";
